add: create borrow feature
Add BorrowController, created with the necessary endpoint
(POST /api/borrows) and the endpoints necessaries to implement
hateoas.
Add BorrowRequestDto, to create the borrow
Add IBorrowService interface and its implementation (the implementation
contains aux method to get the information necessary to be persist)
Change Borrow Entity, remove @TimeStampAnotation, to allow set the date
of finalization.
Change IProductRepository, chenge annotation for repository, to allow
search all types of products (necessary for the aux method to set the
products to persist).